Add configurations to first deploy.
