+++ liveHomeBackend/.gitignore
@@ -114,4 +114,5 @@ dist
 .yarn/unplugged
 .yarn/build-state.yml
 .pnp.*
+.vercel
 store.json
