+++ liveHomeBackend/now.json
@@ -1,0 +1,16 @@
+{
+  "name": "LiveHome",
+  "version": 2,
+  "builds": [{ "src": "index.js", "use": "@now/node-server" }],
+  "routes": [{ "src": "/(.*)", "dest": "index.js" }],
+  "env": {
+    "AUTH_JWT_SECRET": "@lh-jwt-secret",
+    "DB_NAME": "@lh-db-name",
+    "DB_USER": "@lh-db-user",
+    "DB_PASS": "@lh-db-pass",
+    "DB_HOST": "@lh-db-host",
+    "DIALECT": "@lh-db-dialect",
+    "BUCKET_NAME": "@lh-bucket-name",
+    "BASE_URL": "@lh-bucket-base-url"
+  }
+}
