+++ liveHomeBackend/package.json
@@ -4,9 +4,9 @@
   "description": "Web platform that offers properties to the public",
   "main": "index.js",
   "scripts": {
-    "start": "NODE_ENV=production node ./src/index.js",
+    "start": "NODE_ENV=production node ./index.js",
     "test": "mocha",
-    "dev": "nodemon ./src/index.js",
+    "dev": "nodemon ./index.js",
     "lint": "standard",
     "lint:fix": "standard --fix",
     "setup:db": "DEBUG=livehome:* node ./store/setup.js"
@@ -25,6 +25,7 @@
     "bcrypt": "^4.0.1",
     "body-parser": "^1.19.0",
     "chalk": "^4.0.0",
+    "cors": "^2.8.5",
     "debug": "^4.1.1",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
