+++ liveHomeBackend/store/lib/db.js
@@ -5,6 +5,7 @@ let sequelize = null
 
 module.exports = function setupDataBase (config) {
   if (!sequelize) {
+    console.log(sequelize)
     sequelize = new Sequelize(config)
   }
   return sequelize
